ZMLWR-48: refactor all the things!
